+++ core/src/main/java/org/vertexarmy/dsr/core/Root.java
@@ -6,22 +6,21 @@ import lombok.Getter;
 import org.vertexarmy.dsr.core.assets.FontRepository;
 import org.vertexarmy.dsr.core.assets.ShaderRepository;
 import org.vertexarmy.dsr.core.assets.TextureRepository;
-import org.vertexarmy.dsr.core.component.ComponentType;
-import org.vertexarmy.dsr.core.component.InputComponent;
-import org.vertexarmy.dsr.core.component.Node;
-import org.vertexarmy.dsr.core.component.RenderComponent;
+import org.vertexarmy.dsr.core.component.*;
 import org.vertexarmy.dsr.core.systems.InputSystem;
 import org.vertexarmy.dsr.core.systems.RenderSystem;
+import org.vertexarmy.dsr.core.systems.UpdateSystem;
 
 /**
  * created by Alex
  * on 3/21/2015.
  */
 public class Root {
-
     private final InputSystem inputSystem = InputSystem.instance();
 
     private final RenderSystem renderSystem = RenderSystem.instance();
+
+    private final UpdateSystem updateSystem = UpdateSystem.instance();
 
     @Getter
     private UiNode uiNode;
@@ -32,6 +31,7 @@ public class Root {
     public void initialize() {
         Gdx.app.setLogLevel(Application.LOG_DEBUG);
 
+        updateSystem.initialize();
         inputSystem.initialize();
         renderSystem.initialize();
 
@@ -45,29 +45,38 @@ public class Root {
 
     public void addNode(Node node) {
         if (node.hasComponent(ComponentType.RENDER)) {
-            renderSystem.addRenderComponent(getRenderComponent(node));
+            renderSystem.addRenderComponent(getRenderComponentFor(node));
         }
 
         if (node.hasComponent(ComponentType.INPUT)) {
-            inputSystem.addInputComponent(getInputComponent(node));
+            inputSystem.addInputComponent(getInputComponentFor(node));
+        }
+
+        if (node.hasComponent(ComponentType.UPDATE)) {
+            updateSystem.addUpdateComponent(getUpdateComponentFor(node));
         }
     }
 
     public void removeNode(Node node) {
         if (node.hasComponent(ComponentType.RENDER)) {
-            renderSystem.removeRenderComponent(getRenderComponent(node));
+            renderSystem.removeRenderComponent(getRenderComponentFor(node));
         }
 
         if (node.hasComponent(ComponentType.INPUT)) {
-            inputSystem.removeInputComponent(getInputComponent(node));
+            inputSystem.removeInputComponent(getInputComponentFor(node));
+        }
+
+        if (node.hasComponent(ComponentType.UPDATE)) {
+            updateSystem.removeUpdateComponent(getUpdateComponentFor(node));
         }
     }
 
     public void update() {
+        updateSystem.update();
         renderSystem.update();
     }
 
-    private RenderComponent getRenderComponent(Node node) {
+    private RenderComponent getRenderComponentFor(Node node) {
         if (node.hasComponent(ComponentType.RENDER)) {
             return (RenderComponent) node.getComponent(ComponentType.RENDER);
         } else {
@@ -75,9 +84,17 @@ public class Root {
         }
     }
 
-    private InputComponent getInputComponent(Node node) {
+    private InputComponent getInputComponentFor(Node node) {
         if (node.hasComponent(ComponentType.INPUT)) {
             return (InputComponent) node.getComponent(ComponentType.INPUT);
+        } else {
+            return null;
+        }
+    }
+
+    private UpdateComponent getUpdateComponentFor(Node node) {
+        if (node.hasComponent(ComponentType.UPDATE)) {
+            return (UpdateComponent) node.getComponent(ComponentType.UPDATE);
         } else {
             return null;
         }