+++ core/src/main/java/org/vertexarmy/dsr/core/component/ComponentType.java
@@ -6,5 +6,6 @@ package org.vertexarmy.dsr.core.component;
  */
 public enum ComponentType {
     RENDER,
-    INPUT
+    INPUT,
+    UPDATE
 }