+++ core/src/main/java/org/vertexarmy/dsr/core/component/UpdateComponent.java
@@ -1,0 +1,9 @@
+package org.vertexarmy.dsr.core.component;
+
+/**
+ * Created by alex
+ * on 01.04.2015.
+ */
+public interface UpdateComponent {
+    void update();
+}