+++ level_editor/src/org/vertexarmy/dsr/leveleditor/ui/Toolbox.java
@@ -1,12 +1,10 @@
 package org.vertexarmy.dsr.leveleditor.ui;
 
 import com.badlogic.gdx.graphics.g2d.NinePatch;
-import com.badlogic.gdx.scenes.scene2d.Actor;
 import com.badlogic.gdx.scenes.scene2d.ui.ImageButton;
 import com.badlogic.gdx.scenes.scene2d.ui.Skin;
 import com.badlogic.gdx.scenes.scene2d.ui.Table;
 import com.badlogic.gdx.scenes.scene2d.utils.Align;
-import com.badlogic.gdx.scenes.scene2d.utils.ChangeListener;
 import com.badlogic.gdx.scenes.scene2d.utils.NinePatchDrawable;
 import org.vertexarmy.dsr.leveleditor.AssetName;
 
@@ -15,16 +13,19 @@ import org.vertexarmy.dsr.leveleditor.AssetName;
  * on 26.03.2015.
  */
 public class Toolbox extends Table {
-    private final Skin uiSkin;
-
     private final Listener listener;
 
     private ImageButton openLevelButton;
 
     private ImageButton saveLevelButton;
 
+    private ImageButton autoScrollForwardButton;
+
+    private ImageButton pauseAutoScrollButton;
+
+    private ImageButton autoScrollBackwardButton;
+
     public Toolbox(Skin uiSkin, Listener listener) {
-        this.uiSkin = uiSkin;
         this.listener = listener;
         this.setBackground(new NinePatchDrawable(new NinePatch(uiSkin.getRegion("panel"), 1, 1, 2, 1)));
 
@@ -38,6 +39,11 @@ public class Toolbox extends Table {
     private void createComponents() {
         openLevelButton = UIToolkit.createImageButton(AssetName.ICON_FILE_OPEN);
         saveLevelButton = UIToolkit.createImageButton(AssetName.ICON_FILE_SAVE);
+
+        // TODO: Add proper icons here
+        autoScrollForwardButton = UIToolkit.createImageButton(AssetName.ICON_FILE_OPEN);
+        pauseAutoScrollButton = UIToolkit.createImageButton(AssetName.ICON_FILE_OPEN);
+        autoScrollBackwardButton = UIToolkit.createImageButton(AssetName.ICON_FILE_OPEN);
     }
 
     private void layoutComponents() {
@@ -46,20 +52,44 @@ public class Toolbox extends Table {
 
         this.add(openLevelButton);
         this.add(saveLevelButton);
+        this.add(autoScrollBackwardButton);
+        this.add(pauseAutoScrollButton);
+        this.add(autoScrollForwardButton);
     }
 
     private void initListeners() {
-        openLevelButton.addListener(new ChangeListener() {
+        UIToolkit.addActionListener(openLevelButton, new UIToolkit.ActionListener() {
             @Override
-            public void changed(ChangeEvent event, Actor actor) {
+            public void action() {
                 notifyLoadFileRequested();
             }
         });
 
-        saveLevelButton.addListener(new ChangeListener() {
+        UIToolkit.addActionListener(saveLevelButton, new UIToolkit.ActionListener() {
             @Override
-            public void changed(ChangeEvent event, Actor actor) {
+            public void action() {
                 notifySaveFileRequested();
+            }
+        });
+
+        UIToolkit.addActionListener(autoScrollBackwardButton, new UIToolkit.ActionListener() {
+            @Override
+            public void action() {
+                notifyAutoScrollBackward();
+            }
+        });
+
+        UIToolkit.addActionListener(pauseAutoScrollButton, new UIToolkit.ActionListener() {
+            @Override
+            public void action() {
+                notifyPauseAutoScroll();
+            }
+        });
+
+        UIToolkit.addActionListener(autoScrollForwardButton, new UIToolkit.ActionListener() {
+            @Override
+            public void action() {
+                notifyAutoScrollForward();
             }
         });
     }
@@ -76,9 +106,33 @@ public class Toolbox extends Table {
         }
     }
 
+    private void notifyAutoScrollBackward() {
+        if (listener != null) {
+            listener.autoScrollBackwardRequested();
+        }
+    }
+
+    private void notifyPauseAutoScroll() {
+        if (listener != null) {
+            listener.pauseAutoScrollRequested();
+        }
+    }
+
+    private void notifyAutoScrollForward() {
+        if (listener != null) {
+            listener.autoScrollForwardRequested();
+        }
+    }
+
     public interface Listener {
         void loadFileRequested();
 
         void saveFileRequested();
+
+        void autoScrollForwardRequested();
+
+        void pauseAutoScrollRequested();
+
+        void autoScrollBackwardRequested();
     }
 }